rename contenttype.js to content-type.js
